some data changed in properties

+++ AdminSystem/OrdersDataEntry.aspx.cs
@@ -16,9 +16,9 @@ public partial class _1_DataEntry : System.Web.UI.Page
     protected void btnOK_Click(object sender, EventArgs e)
     {
         clsOrder AnOrder = new clsOrder();
-        AnOrder.OrderID = Convert.ToInt16(txtOrderID.Text);
-        AnOrder.CustomerID = Convert.ToInt16(txtCustomerID.Text);
-        AnOrder.ProductID = Convert.ToInt16(txtProductID.Text);
+        AnOrder.OrderID = Convert.ToInt32(txtOrderID.Text);
+        AnOrder.CustomerID = Convert.ToInt32(txtCustomerID.Text);
+        AnOrder.ProductID = Convert.ToInt32(txtProductID.Text);
         AnOrder.DateOfPurchase = Convert.ToDateTime(txtDateOfPurchase.Text);     
         AnOrder.TotalPrice = Convert.ToDecimal (txtTotalPrice.Text);
         AnOrder.ShippingCompany = txtShippingCompany.Text;
